requirements.txt
langchain==0.1.16
langchain-core==0.1.45
langchain-community==0.0.34
langchain-google-genai==1.0.3
faiss-cpu==1.8.0
pandas==2.2.2
python-dotenv==1.0.1
tiktoken==0.6.0